fix(conf) prepare prefix to combine trusted certs
The commit 1e6afa171272a12a1577c7f7e172a5a0cac4ca30 introduced a
regression for a case where we want to enable verification of database
certificates when doing SSL handshake connections. For example, if we
wanted to specify `pg_ssl = on` and `pg_ssl_verify = on` to enable
verification of the postgres certificate, the following error would be
thrown:

```
Error:
/usr/local/share/lua/5.1/kong/cmd/migrations.lua:220: [PostgreSQL error] failed to retrieve PostgreSQL server_version_num: error loading CA locations (No such file or directory)
stack traceback:
[C]: in function 'assert'
/usr/local/share/lua/5.1/kong/cmd/migrations.lua:220: in function 'cmd_exec'
/usr/local/share/lua/5.1/kong/cmd/init.lua:96: in function </usr/local/share/lua/5.1/kong/cmd/init.lua:96>
[C]: in function 'xpcall'
/usr/local/share/lua/5.1/kong/cmd/init.lua:96: in function </usr/local/share/lua/5.1/kong/cmd/init.lua:53>
/usr/local/bin/kong:9: in function 'file_gen'
init_worker_by_lua:48: in function <init_worker_by_lua:46>
[C]: in function 'xpcall'
init_worker_by_lua:55: in function <init_worker_by_lua:53>
```

The aforementioned commit added support for configuring a list of
`lua_ssl_trusted_certificate` and the paths on the list are
concatenated into a single file in a fixed location (`$PREFIX/.
ca_combined`). The configuration used for enabling verification of
database certificates was defined by the new
`lua_ssl_trusted_certificate_combined` property.

Commands such as `kong migrations bootstrap` would fail because they
don't use the `prefix_handler` module. We should call the
`prepare_function` that's inside the `prefix_handler` module in the
`cmd.migrations.lua` module. Also we need to call it earlier in the
`cmd.start.lua` since it tries to initiate a connection to the DB
before preparing the prefix in order for us to generate the
`lua_ssl_trusted_certificate_combined` configuration with the list of
all CAs before the DB connection happens.