feat(pdk) kong.log.set_serialize_value()
Log-type plugins (like file_log, or http_log) use the kong.log.serialize() function to "extract" a table representing the current status of Kong. That table is used in the logs.

This change introduces a new function which allows customizing the table returned by kong.log.serialize(). This can be used by plugins to customize the output of a logging plugin without having to "monkey-patch" it.

Related with #5968