fix(cli) reset kong shm rework on reload
### Summary

The `reset_kong_shm` logic was reworked after finding out multiple issues
with it. One of such issue was reported here: https://github.com/Kong/kong/issues/6464

And there are instructions how to reproduce it. We found further issues when
doing parallel reloads, uploads of new configs and crashing workers at the same
time.

This also removes flushing of non-paged shared dictionaries as that is problematic
with reloads, as there can be old and new workers at the same time. It also prevents
the usage of `:8001/config` during the reload.

### Issues Resolved

Fix #6464