fix(cmd) usage of ffi.C.write could result corrupted .kong_env
### Summary

On my machine (and I have seen this elsewhere too) sometimes a
corrupted `.kong_env` gets generated. This is most probably
because of usage of `ffi.C.write`. It is hard to reproduce,
but it requires multiple reloads and passing `-c` (confs).

I changed the code to use plain Lua, which seemed to mitigate
the issue fully.