chore(ci) use our fork of pgmoon in the makefile (FT-1606) (#2121)
* chore(ci) use our fork of pgmoon in the makefile

Kong’s pgmoon forked repo is public [8] but the features that are on it
are EE only.

There are a few options to use the special pgmoon branch in our kong-ee
builds:

1) In the build scripts, add patches on the fly to the necessary files
after downloading it [1]

Only feasible for simple patches/diffs. For example, [1] was pretty
easy to do, therefore easy to apply as a patch with `sed`. Not only
that, but also becomes a burden to maintain once we bump the upstream
pgmoon dependency in the kong-ee rockspec as it will generate patch
conflicts during the builds. The worst option IMO.

2) Distribute our own fork of pgmoon in a private repository [4] using
the kong-distributions build.yml system [5]

As discussed with @eskibars, it doesn't make sense to take ownership of
the upstream pgmoon project by taking it private as it's not aligned to
our open core, OSS company principles. Discarded.

3) Develop our own fork of pgmoon [2] and then distribute/upload it
publicly to LuaRocks [3]

This will require a major bump by name-changing `pgmoon` to
`pgmoon-kong` in our fork because `kong-lapis` depends on `pgmoon` so
we must prevent namespace issues where kong-lapis' pgmoon dependency
would override our fork. See [9] and [10] for the exact situations
where this issue happened in the past and how we addressed it.

I don't think this is a technically correct thing to do. I don't think
we want `kong-lapis` to use one version of pgmoon (the CE one) and
`kong-ee` itself to use another pgmoon (the EE one). Everything that
depends on pgmoon should use one and only version only. I also don't
think we want to change kong-lapis' pgmoon dependency to `pgmoon-kong`,
otherwise CE would get the EE only feature (CE also uses `kong-lapis`).
Because `kong-lapis` is a fork of the upstream `lapis` project I also
don't think we should create an EE fork of a CE fork just to change the
dependency `pgmoon` to `pgmoon-kong`.

4) The current approach, which was inspired by [6] and [7]

It addresses the issues pointed above in option 3) and it's the
easiest-to-implement option. We are also putting the pgmoon dependency
back in the rockspec since we're overriding it with our own fork in the
Makefile anyway. This is to avoid conflicts in the merges.

By installing pgmoon in the Makefile, it does not break gojira and
other source based installs that do not use kong-distributions.

[1] https://github.com/Kong/kong-distributions/pull/497
[2] https://github.com/leafo/pgmoon/compare/v1.6.0...Kong:2.0.1
[3] https://luarocks.org/modules/kong/pgmoon-mashape
[4] https://github.com/Kong/lua-resty-openapi3-deserializer
[5] https://github.com/Kong/kong-distributions/blob/2.2.0.0/kong-images/build.yml#L97-L101
[6] https://github.com/Kong/kong-ee/commit/1a93a89344d1e88c980f5285269391be707af86b
[7] https://github.com/Kong/kong-distributions/commit/37c65c5b10451328725bab6d9d4e758439db8ad9
[8] https://github.com/Kong/pgmoon
[9] https://github.com/Kong/kong/pull/1582
[10] https://github.com/Kong/pgmoon/compare/1.7.0...2.0.0

* specify KONG_PGMOON_LOCATION to make it consistent with kbt

* split up the installation of kong and pgmoon into two components