fix(response-transformer) do not allow nulls in arrays (#6710)
### Summary

It was reported by @rmlodzikowski-nokia on issue #6707 that
you can configure `response-transformer` with settings that
can lead for runtime error. There were two possible fixes:

1. handle `nulls` on runtime
2. do not allow `nulls` in config

For this PR I picked the `2.` as I think it is better to make
arrays `required` rather than adding `null` handling code in
code. The arrays default to empty array `{}`, so `required`
here does not mean that you need to actively give a value,
it just means that you cannot actively give `null`.

Similar PR will be opened on `request-transformer`.

### Issues Resolved

Fix #6707