fix(license) load license from database (#2158)
* fix(license) load license from database

This will load the license from the database iff the kong.license has
not been set by the normal mechanisms; e.g. environment, default path,
and environment path. If the license is loaded or added to the DB the
license is then reloaded using the lazy load mechanism for the license
utilities.

* Update kong/init.lua

Co-authored-by: Fero <6863207+mikefero@users.noreply.github.com>

Co-authored-by: Rob Serafini <42984308+RobSerafini@users.noreply.github.com>