feat(termination) add a request-echo option
If set the plugin will echo a copy of the received request back to
the client. If an optional trigger is set, then it will only return
the echo if that trigger is defined as a header/query-parameter. If
the trigger is not found, the plugin will not terminate the request.