fix(hybrid) Propagate values for auto fields to data plane
When the control plane data is sent to the data plane the "insert"
context is used which causes auto fields to re-assign their values and
overwrite the value coming from the control plane. The additional
options indicating the update is occurring from a control message allows
for the auto fields to retain their value.