fix(router) use normalized request URI when matching against incoming user requests
Previously Kong's Router have been matching against raw user request
URI, that is, URI without being normalized.

In this commit we change the behavior of Kong Router to always match
against normalized URIs from user. Since Kong Router plays an
important role in generating the URI when proxying to the upstream, it
means that upstream will always see request URIs after normalization as
well.

Note that we can not use the `ngx.var.uri` variable directly in
normalizing the request URI, because it decodes reserved characters as
defined by RFC 3986, which is prohibited by the RFC.