fix(vitals) ensure we only try to save valid cache keys to datastore (FTI-2038) (#2210)
* fix(vitals) ensure only valid cache keys are saved to datastore (FTI-2038)

While introspecting the vitals counters cache keys, I noticed that for
each request to a route with enterprise-proxy-cache enabled and served
out of proxy-cache, there was a duplicate key entry created but with
an empty status code field.

This was preventing the datastore save to work correctly for postgres as
the client couldn't coerce an emtpy status code string into an integer field in the
vitals_codes_by_route table. Therefore those entries never showed up in
vitals for the routes/services or, subsequently in the admin portal graphs.
Only cache Miss data was showing up.

While I couldn't get to the bottom of where those duplicate vitals
counter cache keys were coming from, I did notice that we already had a
pattern for handling this scenario by ensuring the IDX.code field was
added to the required fields when marshaling the cache key into a
datastore insert statement.

My hypothesis is that the duplicate entries without the status code
field are somehow an artifact of a cache phase. Irregardless, they
appear to be ignore-able and the vitals data that subsequently is
preserved in the datastore looks accurate.

With this change, both cache Hits and Misses show up in all the expected
locations for routes/services/workspaces in the admin portal graphs.

Signed-off-by: Jeremy J. Miller <jeremy.miller@konghq.com>