fix(runloop) re-encode special characters before sending to upstream
By default, Nginx's proxy module will escape special characters inside
the upstream request path when generating the request, however, this
behavior is lacking if the request path is overwritten with a variable
at the end of the `proxy_pass` directive (which is what Kong does).

This PR restores that behavior so that the request Kong generates to the
upstream will still follow the HTTP RFC after the router normalization.