fix(mtls-auth) remove CA existence check (FTI-2296)
In FTI-2001 we implemented an existence check for the CA uuid
referenced when the plugin is created. This check is incompatible with
db-less mode. This PR removes the check to unblock anyone utilizing
db-less.

I initially wanted to remove the CA existence check only for db-less
mode and leave it in place for other modes. However, this now seems like
a problematic idea as the plugin behavior becomes inconsistent between db and
db-less modes.

I believe that the long term solution for introducing an existence check
is to add `ca_certificates` into the core DAO. With that, we could then
perform an existence check without needing a db handle to do so.

Signed-off-by: Jeremy J. Miller <jeremy.miller@konghq.com>