fix(mtls-auth) Grab CA from the end of the proof chain instead of the beginning
From what I can gather, a certificate chain starts with the client
certificate, then lists any intermediate authorities, then ends with the
root CA. When searching for the MTLS credential for a request the plugin
instead:

1. Looks for a proof chain using the supplied client certificate and any
   supplied intermediates provided by the client
2. Searches the list of known `ca_certificates` configured on the plugin
3. If it finds a CA that verifies the provided client certificate(s), it
   returns that as a proof chain
  1. It also appears to search the 'store' of known CA certificates for
     any intermediates too, which is convenient
4. Grabs the first certificate out of the proof chain and tries to look
   up MTLS credentials for that certificate

The issue is that the first certificate in the proof chain is the
supplied client certificate. The *last* certificate in the proof chain
is for the CA. This PR changes that logic to correctly grab the CA
certificate from the proof chain.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>