fix(migrations) sort subsystems alphabetically
The `pairs` iteration over `plugin_names` makes running migrations inconsistent - the order in which the subsystems are loaded changes randomly.

This PR stops the random order by sorting the subsystems alphabetically, so at least the order in which migrations are executed is constant.

Sorting things this way is better than having no order at all, but the ideal solution would be sorting plugins topologically (so a plugin having entities depending on another plugin will be loaded later).