fix(openid-connect) do not expose private keys with /openid-connect/jwks
### Summary

In version `1.6.0` the bug was introduced that made clearing of
the private keys not to work on Admin API endpoint `/openid-connect/jwks`,
and thus made private keys exposed on that endpoint.

It is admin api endpoint, but sometimes people proxy there on public
interface or at least from one accessible by their IDP, and thus can
pose a security threat. Especially if IDP is also publicly available
and the client_id is known to attacker that has gained access to private
keys.