fix(rate-limiting-advanced) do not pre-create namespaces on init-worker (#63)
We saw multiple problems with this approach:

* :each() did not respect the global `max_page` and `page_size` settings
  * as a result the database gets high pressure if it has mass entries.
    in extreme cases the db might even crash
* :each() ignores the underlying caching mechanism
* as this is called in the `init_worker` phase, every worker performs a
  database lookup. This increases pressure on the database.

This patch removes direct calls to the database via the
kong.db.plugins entity. On kong's cache_warmup phase
all plugins are already loaded into ngx's shared store.

As a sideffect to this patch the plugin will create namespaces on the
fly. This may result in a slightly increased response time on the first
request.

Signed-off-by: Joshua Schmid <jxs@posteo.de>