feat(file-log) add censoring capabilities
To avoid logging sensitive data, such as API keys, an attribute remover
is added. It enables the possibility of removing specific attributes
from the message that will be logged. A censored_fields field is added
to the schema were the path of the field should be specified separated
by periods.

* add an attribute remover to the file-log plugin
* add a censored_fields field to the file-log plugin schema
* add an unit test for the attribute remover in the file-log test suite