fix(migration) Plugins that depend on consumers should inherit ws_id from consumer
Plugins that depend on a consumer end up with a DB constraint that
requires the consumer and any plugin entity data to exist in the same
workspace. But in the case of a shared consumer it is possible that the
ws_id picked for that consumer does not match the ws_id for the
associated plugin entity from the data in the workspace_entities table.

To avoid hitting this constraint we set the ws_id for each plugin entity
based on the associated consumer, instead of from the workspace_entities
table.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>