fix(balancer) don't assume upstreams don't exist after init phases (#7010)
There may be cases where an upstream is not yet loaded on `init` and
`init_worker` phases (for example, failure on reading the upstreams
list in the `init()` function) and because of that, after these init
phases, we will just always return a `503` to the client with the error
`"name resolution failed"` since the dns-client lib will return a `"dns
server error: 3 name error"` to Kong. So we can't assume the upstream
doesn't exist just because we are past the init phases.

This applies to both the `strict` and `eventual` balancer worker
consistencies.