feat(hmac-auth) add support for the "@request-target" field support which is introduced in later versions of the "Signing HTTP Messages" RFC. closes #3789
Previously the Kong implementation of the RFC included a signing
parameter "request-line" which contains the HTTP request method and
request URI plus the HTTP version number. This is not part of the RFC
draft, and has created issue such as #3789 where changing the HTTP
request version by intermediary proxy caused the signature validation to
fail.

In general, including the HTTP protocol version inside the signature
makes little sense, as the protocol does not identify the resource that
is being accessed, and can be changed by clients/proxies which still
preserving the original semantics of the request.

This PR adopts the "@request-target" field as described inside
https://tools.ietf.org/html/draft-ietf-httpbis-message-signatures-03
which only included the lowercased request method and request URI when
calculating the hash.

Note: we intentionally steered away from the
https://tools.ietf.org/html/draft-cavage-http-signatures-12 RFC
recommendation which uses the "request-target" field instead. This RFC
draft has expired due to long inactivity and the name also has slight
chance of conflict with HTTP headers named "Request-Target". Overall
the Kong hmac-auth plugin did not follow the RFC draft in the beginning,
and this change should not cause any compatibility issues for the users
that are relying on "request-line" today.