fix(plugin_iterator) guard against Lua error that happens inside plugin's "init_worker" handler
Previously when a plugin's "init_worker" handler fails, the rest of the
`init_worker` task are no longer executed. Although this is not
something that should happen in normal situation, there are certain
custom plugins that are not written correctly and failed "init_worker"
causes things like Hybrid mode to break because it's `init_worker` code
no longer executes.

This PR allows the Kong node to finish the rest of the initialization
even in case of plugin's "init_worker" handler error. It should improve
the situation because Kong core's initialization generally does not
depend on the outcome of a particular plugin's "init_worker" handler, and
this improves the stability of the product slightly.

Because failed "init_worker" handler is still considered a serious
event, we use the `stash_init_worker_error` function so the error
message will be printed for each incoming request later. This is the
same behavior as we do for the other initialization failures that could
occur in core.