fix(proxy) the context got shared between different tls keepalive requests
### Summary

@dsteinkopf reported on #7054 that when executing a plugins iterator with two
subsequent tls keepalive request, the wrong plugin is executed. We found out
that we do pass context in ssl phase to plugin iterator and that indeed creates
a connection wide structures in that context, which are then shared between
different keepalive requests, and thus a wrong plugin might be executed.

This commit changes it so that we do not pass context to plugin iterator with
`certificate` phase and thus it won't store data in it.

Huge thanks for @dsteinkopf for reporting this!

### Issues Resolved

Fix #7054