fix(external-plugins): use lowercase for comparing string in error condition
When an instance_id of a plugin external changes, the plugin instance is resetted and retried in the bridge event loop again.
This was failing and none was being retries due to a typo in the comparisson.
This has been identified and tested with heavy traffic when the kong balancer is refreshed and reload into memory, it was not failing with low traffic.