fix(sandbox) allow string.rep in sandboxed environments
Reasoning: the idea behind limiting access to string.rep was
disallowing a single operation from allocating too much memory.

Given that in LuaJIT there are no debug hooks, it is trivial to
implement a loop which allocates a lot of memory on every single
iteration.

Plus, given that the `string` table was global and obtainable via any
sandboxed string, its sandboxing provoked issues on the global state.

This change accepts string.rep in sandboxed environments as a practical
compromise.

Fixes #7062
Fixes #7133