Storing 'clustering' as a singleton so it can be looked up with loaded config by the messaging service
ngx.timer.at also does not allow a class function (I'm probably refering
to that incorrectly, I am new to Lua) to be called because it always
passes 'premature' as the first arg. Refactored the clustering callback
function to accept 'self' as an argument in order to access the instance
configuration.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>