docs(autodoc) patch all navigation files the same way
The previous version of the scripts/autodoc patched the PDK navigation one way (by inserting it into the navigation) and the admin api in another way (by copying a partial file into a special folder in the docs repo, where **another** script needs to be run).

This change simplifies things by inserting both the navigation files using the same function.