tests(fixtures) use dummy timer to ensure Nginx event loop continues after resolver returns
It is determined that the semaphore module of OpenResty contains a bug
that when `:post()` is called right before a timer terminates, the
thread waiting on `:wait()` will not be triggered immediately. This is a
temporary workaround by @javierguerragiraldez until this bug is fixed
inside OpenResty.

See: https://gist.github.com/dndx/8539137ddbdc804a31c1b667158c7efe for a
minimum reproducible example of this issue.