tests(grpc-gateway) new grpc target
Add Go sources, protobuf definition and Make-like process to compile
everything.

First replicate the grpcbin SayHello call.

The protoc compiler, golang-protoc, and golang-grpc-protoc plugins are
required to generate the Go protobuf sourcefiles during development.
Those files are checked in, so if they're not modified (like in CI),
only the Go compiler is required to create the executable.