fix(dao) Add cascade delete entities to EE hooks
FTI-1847: When an entity has a cascade delete relationship with another
entity, and the parent is deleted, 'workspace_entity_counters' are not
decremented.

For example, deleting a consumer with an associated basicauth_credential
will decrement the consumer counter but not the basicauth_credential.

To fix this we need to add the cascade_entities to the hook so EE can
correctly decrement counters for the children entities. The actual
deletes are all handled by the schema layer, it is only the counter
decrementing that is not working.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>