fix(schema) Remove duplicates from constraints list
It appears that Schema.new is called multiple times for each schema.
This behavior was populating the 'constraints' list with duplicate
entries, making it difficult to handle cascade delete constraints in the
dao layer. I added set logic to remove these duplicates.

In the future it would probably make more sense to instead to not
initialize he schema entities multiple times. However, there does need
to be a 2 phase pass on all the schemas to populate the constraints
because they are an inverse relationship.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>