Must iterate _cache[...].constraints with pairs
I modified the internal constraints object to be a Hash instead of an
Array. `ipairs` ignores non-numeric keys so in order to iterate it
correctly we need to use `pairs`. I was already doing this for
non-workspace entities but missed doing it for the workspace special
case in `.get_constraints()`.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>