fix(openid-connect) stop running the plugin when requirements checks fail with anonymous
### Summary

This was first reported in `FTI-2737`. On `1.6.0` the requirements checks
were refactored to anonymous function. That function could call
`response.forbidden(...)` or `response.unauthorized(...)` which normally
stops processing the plugin immediately. Except when `config.anonymous`
is configured, then it just sets anonymous context and headers. When
requirements checks fail the plugin should stop processing any further.

The FTI reported that plugin still did e.g. consumer mapping in that case,
while it should have kept `anonymous` user.

This commit fixes the issue and makes the code to return on failed
requirements checks, even when the plugin is configured for anonymous
access.