feat(termination) add a request-echo option (#6744)
This PR adds 2 features:

    a trigger for the plugin to kick-in
    an option to echo the incoming request back to the client

AD1) if a trigger is set (a header or query-param name), then any request having that header or query parameter will trigger the plugin functionality as configured. If the trigger is undefined in the configuration, the plugin will always kick-in (existing behaviour). If the trigger is defined in the config, but it is not present in the request, the plugin will not do anything, and allow the request to be proxied.

AD2) If set the plugin will echo a copy of the received request back to the client.

Essentially this is a debug aid. The plugin can be added with no impact. Only when the trigger is specified by the client making the request, the plugin will terminate and echo. Another usecase for this would be to turn an existing endpoint into a healthcheck endpoint, only if the trigger is present (configured in the health probe) it responds, otherwise it just proxies.

Alternatively it can be attached to a catch-all route to try and figure out why no route was matched, especially when there are other proxies in between the client and Kong.