fix(declarative) yaml anchors were not properly processed
### Summary

The `lyaml` is used to parse the YAML input. When YAML anchors are used
the library optimizes decoding by reusing referenced table. Our generate
ids procedure did not take that in the account as explained on issue
#7620 by @bforbis. So we ended up inserting the item with same id on SHM
on import. And that could also be seen from admin api where it only listed
the entity once instead of as many times as it has been used and referenced.

### Issues Resolved

Fix #7620