fix(portal) Fix portal template regression in 2.2.1.4
When cascading
https://github.com/Kong/kong-ee/commit/4b7935d5169e237a3c3377bb48e8dc17b1f80e2c
forward from next/2.1.x.x I accidently cascaded an old version of the
`01_initial_files.json` and `01_initial_files.lua` files. This resulted
in the 2.2.1.4, 2.3.3.3, 2.4.1.2, and 2.5.0.1 releases containing
incorrect portal templates.

This commit fixes the version of those portal templates. I verified that
this is the correct version of these files by checking the timestamp at
the beginning of `01_initial_files.json` and it should be 1626976586
(Thursday, July 22, 2021).

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>