fix(keyring) generate keyring if keyring_blob_path is not set
Here the code was attempting to read keyring data locally by calling
`local_keyring_read` on `init()` by doing an early return with:

```
return local_keyring_read(config)
```

The issue was that if `keyring_blob_path` was not configured (the
default), a `nil` value was returned, and this was efectively ignoring
generation of keyring data because the code that follows this early
return is the generation of keyring data from scratch.

We fix this bug by only calling `local_keyring_read` on `init()` if and
only if `keyring_blob_path` is configured in the first place. Otherwise,
proceed with generating keyring data from scratch.