fix(clustering) version compatibility updates for OIDC plugin upgrade
This commit allows for the upgrading of the OIDC plugin to 2.0.x. This
pattern can also be used to remove other plugin field elements which are
not compatible with older dataplanes. It also fixes a bug when there are
no redis configuration items to remove; added redis options for testing.