fix(keyring) do not decrypt on init* phases
If there are data fields keyring-encrypted in the database and we
attempt to decrypt them on init* phases (init or init_worker), we would
get errors like the one in the description of FT-1589.

Since encryption keys are held only in memory, and do not persist past
a restart of the Kong process (e.g., running `kong restart`), we must
export the keyring following its initialization.

Because of this limitation, we won't necessarily have the encryption
keys available during the init* phases because the user may not already
have imported them.

We are trying to solve this by doing an early return of the whole
keyring-encrypted field on a decrypt attempt. This will give the user
some time to import the keys in order for Kong to start recognizing the
decrypted fields. For example, when we try to decrypt plugin fields on
init* phases, we return the encrypted field so that the plugin iterator
does not fail so hard.

Note that at some point (as soon as possible), the user must still
import the keys. Otherwise, the plugin iterator will not properly
execute the plugin that has encrypted fields. Until then, any attempt
to, for example, hit the Admin API `:8001/plugins` endpoint or send a
proxy request to a route where the plugin is applied to, will fail
because Kong won't recognize the encrypted field.

So a natural good improvement here will be to, when kong is restarted
and plugin fields are encrypted, log a crit error message saying that
the keyring needs to be imported ASAP.