tests(healthchecks) mark connection timeouts test as flaky
So I did a diff between the CE `spec/02-integration/05-proxy/10-balancer/
01-healthchecks_spec.lua` file and the EE `spec-ee/02-integration/02-workspaces/
05-healthchecks_spec.lua` file and they’re pretty much the same except some
workspace logic (obviously).

Given this and the fact that this particular test in the CE test suite is not
failing and but EE one is, I say this is another balancer+travis specific issue
and with this I am going ahead and marking it as flaky. The other reason (but a
less convincing one) for marking it as flaky is that most of the tests in that
`spec-ee/02-integration/02-workspaces/05-healthchecks_spec.lua` file are marked
as flaky.