feat(openid-connect) allow redis cluster nodes to be specified by hostname
Whilst redis cluster clients typically require configuration by ip
address, it is possible to bootstrap using a hostname so long as this
resolves to an ip in the cluster. Once bootstrapped, the cluster client
derives its source of truth (i.e. ip addresses) from the cluster quorum
itself.

We keep the field named as `ip` since this is how it is presented to
lua-resty-session and ultimately the cluster client. This is merely a
loosening of validation rules so that hostnames can be taken advantage
of in certain cases (k8s in particular).