fix(core) Deleting an admin should delete the associated consumer regardless of consumer workspace [FTI-2803]
If the request has permissions to delete the admin entity, the
associated consumer should also be deleted. In certain cases this was
not happening. For example, if the admin was created in a 'echo'
workspace and given default workspace super-user permissions, a delete
request against a 'bravo' workspace should succeed.

The code paths associated with this change are complicated because
admins are non-workspaceable but the underlying entities (consumers and
rbac_users) are workspaceable.

This code change ensures that if the request has appropriate permissions
to delete the entity object, the associated consumer and rbac_user are
deleted. This could also be accomplished by setting up a 'CASCADE
DELETE' relationship in the database but that would require inverting
the existing relationship. There is already existing code to handle this
delete, so I opted to fix that code rather than introduce a major
migration.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>