fix(router) properly escape uri captures
### Summary

When we added normalization `kong.tools.uri.normalize`, that function
does percent-decoding on everything, except for the reserved characters.

That means that we basically percent-decode more than just the ranges of
ALPHA (%41–%5A and %61–%7A), DIGIT (%30–%39), hyphen (%2D), period (%2E),
underscore (%5F), or tilde (%7E).

Which is fine for matching (as we run both inputs through it). Though
calling it `normalize` is wrong, but that is another issue.

Because of doing excessive percent-decoding, we endup things leaking from
there. One of the leakeage is uri captures that now contain the captures
based on our raw normalization.

This commit adds escaping to uri captures, and should fix the leak.