feat(api) add a config readiness endpoint
Add a config readiness endpoint. When db="off", it returns a 200 if Kong
has loaded or received a configuration and applied it, or 503 if not.

Compute hashes when loading configuration from the environment, to
distinguish between user-supplied configuration and the default blank
configuration.

Test the config readiness endpoint. This requires actually using the
database strategy for Kong route tests and adding a utility server that
allows tests to directly set SHM values, to circumvent the lack of
normal config load behavior in test environments.