fix(make-dev) bail out if rock installation fails
This fixes the deeper underlying problem related to https://github.com/Kong/kong-pongo/pull/241 .

In those cases the `make dev` overall failed because LuaRocks failed, but it was not signaled. Hence docker assumes the image to be valid and even caches the results of the docker commands. Despite them being faulty.