fix(clustering) catch pl_file.read() failures
Some of this code was written with the assumption that Penlight's
`pl.file.read()` function throws an error, and therefore its output
does not need to be checked. As it turns out, `pl.file.read()` _can_
throw an error, but this behavior is not default (it's also controlled
by a global switch that could be toggled at any time). This silent
failure leads to an unhelpful, cryptic error being thrown when one of
the ssl functions fails on receiving `nil` input:

> [error] 2050#0: init_by_lua error:
>     /usr/local/openresty/lualib/ngx/ssl.lua:378:
>     attempt to get length of local 'pem' (a nil value)

Wrapping this function call in `assert()` is a simple fix to ensure a
more reasonable error is thrown when `pl.file.read()` fails.

See the docs for `pl.utils.on_error()` and `pl.utils.raise()` (called
when `pl.file.read()` fails) for more info on Penlight's error-handling:

https://lunarmodules.github.io/Penlight/libraries/pl.utils.html