fix(conf) check transfer-encoding header presence
The function kong.response.exit() sets the
Content-Length header on the response even if the
Transfer-Encoding header is present. It violates
"RFC 7230 3.3.2". This updates kong/pdk/response.lua to
check the presence of transfer-encoding before setting
content-length header

* don't set Content-Length if the Transfer-Encoding header
is present in the header list provided as the third method argument.
* add a new test case

* return 405 when the method is not supported in the Admin API helpers
* add a new test case in the Admin API test suite

Fix #8083