fix(test) add data plane log checks to version compatibility tests
Since the `sync_status` is set by the control plane, a additional check
of the data plane logs needs to be performed in order to account for any
bad configurations sent to the data plane. If any error messages exist
that match the bad configuration pattern, they will be output to the
terminal and force a failed exit for Jenkins failure.

- Build and create a snapshot for sequential pipelines
- Update default data plane version
- Do not skip enterprise plugins
- Fix Jenkinsfile linting issues