fix(exit-transformer) avoid manually utilising the plugins iterator
This plugin is a little curious, as it runs on an `exit` hook rather
than in one of the plugin phases. During `exit` it needs to locate the
most relevant plugin instance config. Note that this may be a global
instance in the case of an unmatched route.

Previouisly this was acheived by manually calling the plugins iterator,
which stopped working as expected due to performance improvements in
2.7. Calling the iterator in this way caused later plugins to stop
  working - i.e. it intefered with the runloop.

This commit reworks the implementation to:

- Use our access handler to pass the plugin instance config into the
  request context.

- In the case where we have no config (i.e. an unmatched route) it uses
  the cache to try and locate a global instance for workspace.