fix(datadog) default value for metrics specified twice
### Summary

I found out that datadog schema specifies default for metrics in
two levels. Once in config field level and second in metrics level.
The config field level causes validation errors in schema loading
but those errors are ignored by our loader for some reason. Still
I think this is a bug and the default should not be there twice.