feat(datadog) allow adding tags to all Datadog metrics
Without this, the only way to add tags to every metric emitted by the
Datadog plugin is to completely specify the entire metric list, which
is tedious, error-prone, and not forward-compatible. This adds a
feature that lets a user specify a list of tags to add to every metric
emitted by the Datadog plugin.