chore(*) Adding a Kong Portal manifest file for generating copyrights
Originally I was going to add this manifest to kong-portal-templates and
update kong-portal-tools to copy it over when it copied
01_initial_files.json.

But this manifest file is currently generated manually since there is
not automated dependency bumping in the kong-portal projects.
kong-portal-tools also does not have automation for opening the PR when
copying 01_initial_files.json.

I decided to make a 1 time check-in of the manually generated portal
manifest. When we make automation on the kong-portal side to manage
dependencies, we can introduce automation to also overwrite the manifest
file stored in kong-ee.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>