chore(prometheus) increment plugin version
This commit increments the Prometheus plugin handler version. There are
no functional/code changes here, but there will be changes in Kong EE
for this version. It's unfortunate, but it's slightly less problematic
than letting the versions get out-of-sync between the two codebases.

I'm unsure if the rockspec file is needed anymore, as it was already one
version behind the version in the plugin handler file, but I've updated
it to match these changes anyway.