feat(vaults) add environment variable vault implementation
### Summary

The `env` vault can resolve references from environment variables.

The environment variable can be a any string or JSON string if it
is referenced by a key. We also support `_PREVIOUS` suffix to load
the previous version.

Here is a reference string for process secret using `env` vault:
```
export MY_SECRET=foo
```
```
{vault://env/my_secret}
```

And here is how it can be used with a key:
```
export MY_SECRET='{"username":"foo","password":"bar"}'
```
```
{vault://env/my_secret/username}
{vault://env/my_secret/password}
```