fix(pdk) change detection of process/config secrets
### Summary

Previously we used `ngx.IS_CLI` to determine whether we want config secret or
a process secret, that is:
- should we look database for entity prefix
- should we look our vault implementations by name

We decided (with Joshua) that talking about process secrets / config
secrets is confusing to users, so I changed the detection algorithm to:

We look for implementation by name when:
1. name is one of the bundled vaults or specified in ´KONG_VAULTS=...`
2. OR kong and kong.db is uninitialized

Otherwise, we look for database prefix.

This PR also disallows configuring Vault entity prefix with value
that is a name of one of the bundled vaults or specified in `KONG_VAULTS`.

I also added admin API tests with this.